// src/hooks/types.ts
export declare namespace HooksWallet {
  export interface Accounts {
    address: string;
    alianName: string;
    type: string;
    brandName: string;
  }

  export interface CurrentAccount {
    address: string;
    alianName: string;
    type: string;
    brandName: string;
    shortAddress: string;
    originalAddress: string;
  }
}